Refactored CommandParser and renamed it to ArgumetsParser
